+++ CHANGELOG.rst
@@ -1,3 +1,10 @@
+**[Unreleased]**
+
+*Added*
+
+- `conkit.io` routines now accept keyword arguments
+- SAINT2 and ROSETTA distance restraints can now be written, `format` keywords are `saint2` and `rosetta`
+
 **[0.10.2]**
 
 - `MANIFEST.ini` file required by PyPi