+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
 
-**[Unreleased]**
+**[0.11.1]**
 
 *Changed*